apps/worker/requirements.txt
redis==5.0.1
rq==1.16.1
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
PyMuPDF==1.23.8
python-docx==1.1.0
pillow==10.2.0
minio==7.2.3
